+++ package.json
@@ -29,6 +29,7 @@
     "googleapis": "^47.0.0",
     "lodash.template": "^4.5.0",
     "mailgun-js": "^0.22.0",
+    "multer": "^1.4.2",
     "pug": "^2.0.4"
   },
   "release-it": {
